Loads of stuff! Finished for class
Did the whole entire info area. Added a new enemy type. Refined the physics. Fixed some bugs. Made the spawning system better. Scratching now summons a random ball anywhere, instead of an 8 ball in the middle. The game is basically complete now. I would have liked to add more but this is it for my class.